+++ src/components/courierInboxVue.ts
@@ -1,6 +1,7 @@
-import { h, type PropType, defineComponent } from 'vue-demi';
+import { h, type PropType, defineComponent, watch } from 'vue-demi';
 
-import type { InboxPlacement, InboxTheme, Brand, Labels, InboxTrigger } from '../types/inbox';
+import type { InboxPlacement, InboxTheme, Brand, Labels, InboxTrigger, OnEvent, InboxProps } from '../types/inbox';
+import { Courier } from '../ts/courier';
 
 type IsoFormatter = (isoDate: string) => string;
 
@@ -9,20 +10,37 @@ export default defineComponent({
     tenantId: String,
     brand: Object as PropType<Brand>,
     className: String,
-    openLinksInNewTab: Boolean,
-    renderActionsAsButtons: Boolean,
-    showUnreadMessageCount: Boolean,
+    defaultIcon: String,
     from: Number,
-    views: Array as PropType<{id: string, label: string, params: any}[]>,
+    isOpen: { type: Boolean, default: undefined },
+    views: Array as PropType<{ id: string; label: string; params: any }[]>,
     formatDate: Function as PropType<IsoFormatter>,
+    appendTo: String,
     labels: Object as PropType<Labels>,
-    defaultIcon: { type: [Boolean, String] },
+    onEvent: Object as PropType<OnEvent>,
+    openLinksInNewTab: { type: Boolean, default: undefined },
     placement: Object as PropType<InboxPlacement>,
+    showUnreadMessageCount: { type: Boolean, default: undefined },
     theme: Object as PropType<InboxTheme>,
     title: String,
     trigger: Object as PropType<InboxTrigger>
+    // The following cannot be easily converted from React.FunctionComponent to a vue equivalent
+    // renderContainer:
+    // renderBell:
+    // renderFooter:
+    // renderPin:
+    // renderIcon:
+    // renderMessage:
+    // renderNoMessages
   },
   setup(props) {
-    return () => h('courier-inbox', { ...props });
+    /* 
+    https://www.courier.com/docs/platform/inbox/web/javascript-sdk/
+    https://github.com/trycourier/courier-react/tree/main/packages/components#updating-component-configs
+    components have centralised properties when using the embedded version, so we need to make sure
+    props are converted from vue prop semantics into global updates 
+    */
+    watch(props, p => Courier.updateInbox(p as any as InboxProps), { immediate: true, flush: 'pre' });
+    return () => h('courier-inbox', { 'append-to': '#inbox-container' });
   }
 });