+++ src/types/toast.ts
@@ -1,4 +1,9 @@
-import { CSSObject } from "styled-components";
+import { CSSObject } from 'styled-components';
+import { Brand } from './inbox';
+
+export type ToastPosition = 'top-left' | 'top-right' | 'top-center' | 'bottom-left' | 'bottom-right' | 'bottom-center';
+export type ToastRole = 'alert' | 'status';
+export type ToastTransition = 'bounce' | 'slide' | 'zoom';
 
 export interface ToastTheme {
   body?: CSSObject;
@@ -16,6 +21,16 @@ export interface ToastTheme {
   progressBar?: CSSObject;
 }
 
-export type ToastPosition = 'top-left' | 'top-right' | 'top-center' | 'bottom-left' | 'bottom-right' | 'bottom-center';
-export type ToastRole = 'alert' | 'status';
-export type ToastTransition = 'bounce' | 'slide' | 'zoom';
+export interface ToastProps {
+  autoClose?: false | number;
+  brand?: Brand;
+  defaultIcon?: string | false;
+  hideProgressBar?: boolean;
+  // onClick?: React.MouseEventHandler<Element>;
+  openLinksInNewTab?: boolean;
+  position?: ToastPosition;
+  role?: string;
+  theme?: ToastTheme;
+  transition?: string;
+  // renderMessage?: React.FunctionComponent<IInboxMessagePreview>;
+}